requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cryptography==40.0.2
defusedxml==0.7.1
Django==4.2
django-allauth==0.54.0
django-filter==23.1
idna==3.4
kombu==5.2.4
oauthlib==3.2.2
Pillow==9.5.0
prompt-toolkit==3.0.38
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2023.3
redis==4.5.5
requests==2.29.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
tzdata==2023.3
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
